chore: 피드백 타임라인 수정

+++ BE/src/main/java/com/ssafy/db/mapping/FeedbackResMapping.java
@@ -2,7 +2,7 @@ package com.ssafy.db.mapping;
 
 public interface FeedbackResMapping {
 
-    int getTimeline();
+    float getTimeline();
     String getFeedbackType();
     String getFeedbackContent();
 }
